+++ integration-tests/src/test/java/com/transferwise/tasks/testapp/TaskDataIntTest.java
@@ -137,8 +137,4 @@ public class TaskDataIntTest extends BaseIntTest {
     assertThat(oldData).isEqualTo("");
   }
 
-  @Test
-  void testTransactionalInsert(){
-
-  }
 }